requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
mysql-connector-python==8.2.0
pillow==10.1.0
